Resolved int vs uint32_t return type conflict

+++ problem2_starter/limited_size_bag.tpp
@@ -22,7 +22,7 @@ template <typename T>
 bool LimitedSizeBag<T>::remove(const T &item)
 {
   bool found = false;
-  for (uint8_t i = 0; i < current_size; i++)
+  for (int i = 0; i < current_size; i++)
   {
     if (items[i] == item) // Check for first match
     {
@@ -48,7 +48,7 @@ bool LimitedSizeBag<T>::isEmpty() const
 }
 
 template <typename T>
-uint32_t LimitedSizeBag<T>::getCurrentSize() const
+int LimitedSizeBag<T>::getCurrentSize() const
 {
   return current_size;
 }
@@ -56,7 +56,7 @@ uint32_t LimitedSizeBag<T>::getCurrentSize() const
 template <typename T>
 bool LimitedSizeBag<T>::contains(const T &item) const
 {
-  for (uint8_t i = 0; i < current_size; i++)
+  for (int i = 0; i < current_size; i++)
   {
     if (items[i] == item) // Check for match
     {
@@ -69,7 +69,7 @@ bool LimitedSizeBag<T>::contains(const T &item) const
 template <typename T>
 void LimitedSizeBag<T>::clear()
 {
-  for (uint8_t i = 0; i < current_size; i++)
+  for (int i = 0; i < current_size; i++)
   {
     items[i] = T(); // Reset to default value
   }
@@ -77,7 +77,7 @@ void LimitedSizeBag<T>::clear()
 }
 
 template <typename T>
-uint32_t LimitedSizeBag<T>::getFrequencyOf(const T &item) const
+int LimitedSizeBag<T>::getFrequencyOf(const T &item) const
 {
   uint32_t count = 0;
   for (uint8_t i = 0; i < current_size; i++)
@@ -97,7 +97,7 @@ bool LimitedSizeBag<T>::operator==(const AbstractBag<T> &other) const
   {
     return false;
   }
-  for (uint8_t i = 0; i < current_size; i++)
+  for (int i = 0; i < current_size; i++)
   {
     if (this->getFrequencyOf(items[i]) != other.getFrequencyOf(items[i])) // Compare amounts of each item
     {
